third_party/blink/web_tests/platform/generic/fast/selectors/querySelector-leftmost-selector-matches-ancestor-expected.txt

Test for crbug.com/304069: querySelectorAll regression.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rootNode.querySelectorAll('#body div').length is 6
PASS rootNode.querySelectorAll('#body').length is 0
PASS rootNode.querySelectorAll('div + #rootNode div').length is 6
PASS rootNode.querySelectorAll('.sibling + .rootNode div').length is 6
PASS rootNode.querySelectorAll('.sibling ~ .rootNode div').length is 6
PASS rootNode.querySelectorAll('.sibling .rootNode div').length is 0
PASS rootNode.querySelectorAll('.sibling div').length is 0
PASS rootNode.querySelectorAll('.body .rootNode div').length is 6
PASS rootNode.querySelectorAll('.body .sibling + .rootNode div').length is 6
PASS rootNode.querySelectorAll('.body .parentSibling + .parent div').length is 6
PASS rootNode.querySelectorAll('.body .parentSibling ~ .parent div').length is 6
PASS rootNode.querySelectorAll('.body div').length is 6
PASS rootNode.querySelectorAll('.body').length is 0
PASS rootNode.querySelectorAll('body div').length is 6
PASS rootNode.querySelectorAll('body').length is 0
PASS rootNode.querySelector('#body') is null
PASS rootNode.querySelector('.body') is null
PASS rootNode.querySelector('body') is null
PASS successfullyParsed is true

TEST COMPLETE
